+++ examples/client/src/main/scala/org/patricknoir/kafka/examples/client/SimpleRSClient.scala
@@ -5,13 +5,13 @@ import akka.util.Timeout
 import cats.data.Xor
 import org.patricknoir.kafka.reactive.client.KafkaReactiveClient
 import org.patricknoir.kafka.reactive.client.config.KafkaRClientSettings
-import scala.concurrent.{Await, Future}
+import scala.concurrent.{ Await, Future }
 import scala.concurrent.duration._
 import scala.io.StdIn
 
 /**
-  * Created by patrick on 09/08/2016.
-  */
+ * Created by patrick on 09/08/2016.
+ */
 object SimpleRSClient extends App {
 
   implicit val system = ActorSystem("ReactiveClient")
@@ -24,7 +24,8 @@ object SimpleRSClient extends App {
   do {
     val response: Future[Error Xor String] = client.request[String, String]("kafka:simple/echo", "hello world!")
 
-    response.onComplete{r => println(r)
+    response.onComplete { r =>
+      println(r)
       println("press e + enter to exit, otherwise enter to repeat the request!")
     }
