+++ kafka-reactive-system/src/main/scala/org/patricknoir/kafka/reactive/client/actors/KafkaRClientActor.scala
@@ -7,7 +7,7 @@ import akka.actor._
 import akka.event.LoggingReceive
 import akka.util.Timeout
 import org.patricknoir.kafka.reactive.client.actors.KafkaRClientActor.KafkaRequest
-import org.patricknoir.kafka.reactive.common.ReactiveDeserializer
+import org.patricknoir.kafka.reactive.common.{ Destination, ReactiveDeserializer }
 import org.patricknoir.kafka.reactive.ex.{ ConsumerException, ProducerException }
 
 import scala.concurrent.duration._
@@ -39,18 +39,6 @@ class KafkaRClientActor(producerProps: Props, consumerProps: Props) extends Acto
 }
 
 object KafkaRClientActor {
-
-  case class Destination(medium: String, topic: String, serviceId: String)
-  object Destination {
-    def unapply(destination: String): Option[(String, String, String)] = Try {
-      val mediumAndTopic = destination.split(":")
-      val medium = mediumAndTopic(0)
-      val topicAndRoutes = mediumAndTopic(1).split("/")
-      val topic = topicAndRoutes(0)
-      val route = topicAndRoutes.drop(1).mkString("/")
-      (medium, topic, route)
-    }.toOption
-  }
 
   case class KafkaRequest(destination: Destination, payload: String, timeout: Timeout, replyTo: String, decoder: ReactiveDeserializer[_])
