+++ kafka-reactive-system/src/test/scala/org/patricknoir/kafka/reactive/server/ReactiveSystemSpec.scala
@@ -4,16 +4,15 @@ import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
 import akka.stream.scaladsl._
 import akka.testkit.TestKit
-import org.patricknoir.kafka.reactive.server.dsl._
+import io.circe.syntax._
 import org.patricknoir.kafka.reactive.client.actors.KafkaConsumerActor.{ KafkaResponseEnvelope, KafkaResponseStatusCode }
 import org.patricknoir.kafka.reactive.client.actors.KafkaProducerActor.KafkaRequestEnvelope
+import org.patricknoir.kafka.reactive.common.Destination
+import org.patricknoir.kafka.reactive.server.dsl._
 import org.specs2.SpecificationLike
-import io.circe.syntax._
-import org.patricknoir.kafka.reactive.client.actors.KafkaRClientActor.Destination
 
-import scala.concurrent.{ Await, Future }
 import scala.concurrent.duration.Duration
-import scala.util.Try
+import scala.concurrent.{ Await, Future }
 
 /**
  * Created by patrick on 13/07/2016.
